Move test to more suitable location
